+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.7.0 [unreleased]
 
+1. [#69](https://github.com/influxdata/influxdb-client-csharp/pull/69): Write buffer uses correct flush interval and batch size under heavy load
+
 ## 1.6.0 [2020-03-13]
 
 ### Features